+++ create-image.sh
@@ -20,4 +20,5 @@ native-image --verbose \
     --initialize-at-run-time=org.postgresql.sspi.NTDSAPI \
     --initialize-at-run-time=com.sun.jna.platform.win32.Secur32 \
     --initialize-at-run-time=com.sun.jna.platform.win32.Kernel32 \
+    --initialize-at-run-time='java.lang.Math$RandomNumberGeneratorHolder' \
     -jar "$uberjar"