+++ samples/db/clj/common/util.clj
@@ -1,0 +1,5 @@
+(ns common.util)
+
+(defn print-result
+  [result]
+  (println "Rows processed:" result))