+++ samples/db/migrations/20201106155531-programmatic-third.down.clj
@@ -1,0 +1,7 @@
+(require '[next.jdbc :as jdbc]
+         '[common.util :refer [print-result]])
+
+(println "Programmatic DOWN")
+
+(print-result
+  (:next.jdbc/update-count (jdbc/execute-one! (get-connection) ["delete from bar"])))