+++ samples/db/migrations/20201106155531-programmatic-third.up.clj
@@ -1,0 +1,11 @@
+(require '[next.jdbc.sql :as sql]
+         '[common.util :as util])
+
+(defn prepare-ids [n]
+  (mapv vector (repeatedly n random-uuid)))
+
+(println "Programmatic UP")
+
+(let [data (prepare-ids 5)]
+  (util/print-result
+    (count (sql/insert-multi! (get-connection) :bar [:id] data))))